requirements.txt
aiogram==3.1.1
feedparser==6.0.10
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
pytz==2023.3
